+++ src/Router/Executor/Exception/ParameterValueNotFound.php
@@ -6,10 +6,10 @@ namespace ExtendsSoftware\ExaPHP\Router\Executor\Exception;
 use ExtendsSoftware\ExaPHP\Router\Executor\ExecutorException;
 use InvalidArgumentException;
 
-class ParameterNotFound extends InvalidArgumentException implements ExecutorException
+class ParameterValueNotFound extends InvalidArgumentException implements ExecutorException
 {
     /**
-     * ParameterNotFound constructor.
+     * ParameterValueNotFound constructor.
      *
      * @param string $name
      */
@@ -17,8 +17,8 @@ class ParameterNotFound extends InvalidArgumentException implements ExecutorExce
     {
         parent::__construct(
             sprintf(
-                'Parameter name "%s" can not be found in route match parameters and has no default value or ' .
-                'allows null.',
+                'Value for parameter "%s" can not be found in route match parameters or request attributes and ' .
+                'has no default value or allows null.',
                 $name
             )
         );